refactor: type registry, unified issues() signature, Project aggregate
- Document._REGISTRY + register() decorator: subclasses self-register,
  Document.parse() uses registry lookup — no lazy imports, no subclass
  knowledge in the base class
- wos/__init__.py imports all type modules to trigger registration
- issues() on every class now accepts **_ to absorb unknown kwargs;
  validate_file/validate_project call doc.issues() uniformly — no isinstance
- wos/project.py: Project aggregate with validate(), check_project_files(),
  check_indexes(); validate_file moved here
- wos/validators.py: thin backwards-compatible shim over project.py
